+++ src/Payum/Core/Bridge/Doctrine/Resources/mapping/Payment.mongodb.xml
@@ -4,7 +4,7 @@
                         xsi:schemaLocation="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping
                         http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping.xsd">
 
-    <mapped-superclass name="Payum\Core\Model\Order">
+    <mapped-superclass name="Payum\Core\Model\Payment">
 
         <field name="number" type="string" />
