--- src/Payum/Core/Bridge/Propel/Model/Order.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Payum\Core\Bridge\Propel\Model;
-
-use om\BaseOrder;
-use Payum\Core\Model\OrderInterface;
-
-class Order extends BaseOrder implements OrderInterface
-{
-}
