+++ src/Payum/Core/Bridge/Propel/Model/Payment.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Payum\Core\Bridge\Propel\Model;
+
+use om\BasePayment;
+use Payum\Core\Model\PaymentInterface;
+
+class Payment extends BasePayment implements PaymentInterface
+{
+}
