--- src/Payum/Core/Bridge/Propel2/Model/Order.php
@@ -1,8 +1,0 @@
-<?php
-namespace Payum\Core\Bridge\Propel2\Model;
-
-use Payum\Core\Bridge\Propel2\Model\Base\Order as BaseOrder;
-
-class Order extends BaseOrder
-{
-}
