+++ src/Payum/Core/Bridge/Propel2/Model/Payment.php
@@ -1,0 +1,8 @@
+<?php
+namespace Payum\Core\Bridge\Propel2\Model;
+
+use Payum\Core\Bridge\Propel2\Model\Base\Payment as BaseOrder;
+
+class Payment extends BasePayment
+{
+}
